android: Remove history fragment
