--- support/android/apk/servoapp/src/main/res/layout/activity_history.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="org.servo.servoshell.HistoryActivity">
-
-    <androidx.fragment.app.FragmentContainerView
-        android:id="@+id/history_container"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:name="org.servo.servoshell.HistoryFragment" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
